% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{anova_SPSS}
\alias{anova_SPSS}
\title{Complete SPSS-like output for ANOVA}
\usage{
anova_SPSS(dataT, dv, wid, within = NULL, within_full = NULL,
  within_covariates = NULL, between = NULL, between_covariates = NULL,
  type = 3, genEs = F, DoFd = 2, Fd = 2, pd = 3, pc = 0.001,
  e2d = 3, e2c = 0.001, epd = 3, epc = 0.001, td = 2, dd = 3,
  DescDigits = 2, Eng = T, Std = T, Unit = NULL, printOut = T,
  headerL = 1)
}
\arguments{
\item{dataT}{Data table to input.}

\item{dv}{Name of the column in data that contains the dependent variable. Values in this column must be numeric.}

\item{wid}{Name of the column in data that contains the variable specifying the case/Ss identifier. This should be a unique value per case/Ss.}

\item{within}{Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{within_full}{Same as within, but intended to specify the full within-Ss design in cases where the data have not already been collapsed to means per condition specified by within and when within only specifies a subset of the full design.}

\item{within_covariates}{Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss and are to serve as covariates in the analysis. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{between}{Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{between_covariates}{Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss and are to serve as covariates in the analysis. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{type}{Numeric value (either 1, 2 or 3) specifying the Sums of Squares “type” to employ when data are unbalanced (eg. when group sizes differ). type = 2 is the default because this will yield identical ANOVA results as type = 1 when data are balanced but type = 2 will additionally yield various assumption tests where appropriate. When data are unbalanced, users are warned that they should give special consideration to the value of type. type=3 will emulate the approach taken by popular commercial statistics packages like SAS and SPSS, but users are warned that this approach is not without criticism.}

\item{genEs}{Whether to print the generalized eta squared or not}

\item{DoFd}{decimal place for degrees of freedom}

\item{Fd}{decimal place for F stat}

\item{pd}{decimal place for p value}

\item{pc}{cutoff value for p value. (i.e. lower limit)}

\item{e2d}{decimal place for eta squared}

\item{e2c}{cutoff value for eta squared. (i.e. lower limit)}

\item{epd}{decimal place for epsilon}

\item{epc}{cutoff value for epsilon}

\item{td}{decimal place for the t statistics}

\item{dd}{decimal place for Cohen's d}

\item{DescDigits}{decimal place for the descriptive statistics}

\item{Eng}{whether the output is in English (true) or French (false)}

\item{Std}{whether to report the standard deviation (true) or the standard error (false)}

\item{Unit}{measurement unit of the DV}

\item{printOut}{if True, print the results in the HTML format.}

\item{headerL}{header level for the printout.}
}
\value{
a list containing the results from the ezANOVA call as well as data tables that can be printed by kable
}
\description{
This function prints and returns the complete results from an ANOVA (ANOVA results, descriptive stats and test of simple effects)
}
