% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\docType{data}
\name{donn_ag}
\alias{donn_ag}
\title{Example data for muCoVL}
\format{data.table with 711 observations of 9 variables.}
\usage{
data("donn_ag")
}
\description{
A data set used for examples.
}
\details{
\itemize{
\item sub. Participant code.
\item Trial_Block. factor indicating whether the participant is in the Block or Trail condition
\item Group. Language group. French or Italian
\item dist. Number distribution. 3 for (1,2,3), 6 for (4,5,6) and 9 for (7,8,9)
\item dur. Target duration. 3000, 7000 or 10000 ms.
\item relE. Relative error.
\item cov. Coefficient of Variation.
\item CEr. Constant Error.
\item AEr. Absolute Error.
}
}
