% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{effSizeVLDT}
\alias{effSizeVLDT}
\title{Calculate basic statistics for a data.table}
\usage{
effSizeVLDT(dataT, facS, otherF = NULL, DV, subF, alpha = 0.05, Eng = T,
  Std = T, Unit = "", digits = 2)
}
\arguments{
\item{dataT}{data.table}

\item{facS}{Name of the factors on which to calculate the descriptive statistics. Can be a character vector or a single comma separated character value.}

\item{otherF}{Name of the other factors to aggregate.}

\item{DV}{Name of the dependant variable.}

\item{subF}{Name of the subject factor.}

\item{alpha}{Confidence level.}

\item{Eng}{whether the output is in English (true) or French (false)}

\item{Std}{whether to report the standard deviation (true) or the standard error (false)}

\item{Unit}{measurement unit of the DV}

\item{digits}{number of rounding digits.}
}
\value{
a data.table with the following columns
\itemize{
\item Mean: Sample mean.
\item StD: Sample standard deviation.
\item StE: Sample standard error.
\item NN: Sample size.
\item CIlow: Lower bound of the confidence interval.
\item CIhigh: Higher bound of the confidence interval.
}
}
\description{
This function calculates the descriptive statistics of a data table.
}
\examples{
data("donn_ag",package="muCoVL")
out<- effSizeVLDT(dataT=donn_ag,facS="Trial_Block,Group,dur",otherF="dist",DV="relE",subF="sub")
}
