% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{ezA_VL}
\alias{ezA_VL}
\title{Augment ezANOVA output}
\usage{
ezA_VL(ezOut, digits = 2, genEs = F)
}
\arguments{
\item{ezOut}{ezANOVA output.}

\item{digits}{significant digits.}

\item{genEs}{Whether to print the generalized eta squared or not}
}
\value{
augmented list. The txxthtml is a printable an HTML table. Use in conjunction with xtable.
}
\description{
This function augments the output from an ezANOVA for easier pasting into a manuscript.
}
\examples{
data("donn_ag",package="muCoVL")
donn_ag[,durF:=factor(dur)]
outt<-ezA_VL(ezANOVA(data = donn_ag,dv = .(relE),wid = .(sub),within = .(dist,durF),between = .(Trial_Block,Group),type = 3,return_aov = T))
}
