% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{ezA_VL_Direct}
\alias{ezA_VL_Direct}
\title{Augment ezANOVA output}
\usage{
ezA_VL_Direct(dataT, dv, wid, within = NULL, within_full = NULL,
  within_covariates = NULL, between = NULL, between_covariates = NULL,
  type = 3, digits = 2, genEs = F, DoFd = 2, Fd = 2, pd = 3,
  pc = 0.001, e2d = 3, e2c = 0.001, epd = 3, epc = 0.001)
}
\arguments{
\item{dataT}{Data table to input.}

\item{dv}{Name of the column in data that contains the dependent variable. Values in this column must be numeric.}

\item{wid}{Name of the column in data that contains the variable specifying the case/Ss identifier. This should be a unique value per case/Ss.}

\item{within}{Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{within_full}{Same as within, but intended to specify the full within-Ss design in cases where the data have not already been collapsed to means per condition specified by within and when within only specifies a subset of the full design.}

\item{within_covariates}{Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss and are to serve as covariates in the analysis. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{between}{Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{between_covariates}{Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss and are to serve as covariates in the analysis. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{type}{Numeric value (either 1, 2 or 3) specifying the Sums of Squares “type” to employ when data are unbalanced (eg. when group sizes differ). type = 2 is the default because this will yield identical ANOVA results as type = 1 when data are balanced but type = 2 will additionally yield various assumption tests where appropriate. When data are unbalanced, users are warned that they should give special consideration to the value of type. type=3 will emulate the approach taken by popular commercial statistics packages like SAS and SPSS, but users are warned that this approach is not without criticism.}

\item{digits}{significant digits. Deprecated.}

\item{genEs}{Whether to print the generalized eta squared or not}

\item{DoFd}{decimal place for degrees of freedom}

\item{Fd}{decimal place for F stat}

\item{pd}{decimal place for p value}

\item{pc}{cutoff value for p value. (i.e. lower limit)}

\item{e2d}{decimal place for eta squared}

\item{e2c}{cutoff value for eta squared. (i.e. lower limit)}

\item{epd}{decimal place for epsilon}

\item{epc}{cutoff value for epsilon}
}
\value{
augmented list. The txxthtml is a printable an HTML table. Use in conjunction with xtable.
}
\description{
This function augments the output from an ezANOVA for easier pasting into a manuscript. Unlike ezA_VL, it invokes ezANOVA directly
}
\examples{
data("donn_ag",package="muCoVL")
donn_ag[,durF:=factor(dur)]
outt<-ezA_VL_Direct(data = donn_ag,dv = .(relE),wid = .(sub),within = .(dist,durF),between = .(Trial_Block,Group))
}
