% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{graDescVL}
\alias{graDescVL}
\title{Prints a graph of descriptive statistics}
\usage{
graDescVL(dataT, xx, groupF = NULL, facetF = NULL, otherF = NULL, DV,
  subF, alpha = 0.05, dWidth = 0.5, eWidth = 0.5, pSize = 3,
  pFill = "white")
}
\arguments{
\item{dataT}{Data.table containing the data.}

\item{xx}{Name of the factor to put on the x axis of the graph. Must be singular.}

\item{groupF}{Name of the factor to use as the grouping factor of the graph. Must be singular}

\item{facetF}{Name of the factor(s) to use as faceting. There can be more than one. Either a single comma separated character or a character vector.}

\item{otherF}{Name of the other factors to aggregate.}

\item{DV}{Name of the dependant variable.}

\item{subF}{Name of the subject factor.}

\item{alpha}{Confidence level.}

\item{dWidth}{Dodge width.}

\item{eWidth}{Error bar width.}

\item{pSize}{Point size.}

\item{pFill}{Fill colour.}
}
\value{
A graph.
}
\description{
Prints a graph of the descriptive statistics of the data in a data.table
}
\examples{
data("donn_ag",package="muCoVL")
dataT<-donn_ag
xx<-"dur"
groupF<-"Group"
facetF<-"Trial_Block"
otherF<-"dist"
DV<-"relE"
subF<-"sub"
result<-graDescVL(dataT = dataT,xx = xx,groupF = groupF,facetF = facetF,otherF = otherF,DV = DV,subF = subF,eWidth = 1000,dWidth = 1000)
}
