% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{pairedCVLDT}
\alias{pairedCVLDT}
\title{Paired comparisons on data.tables}
\usage{
pairedCVLDT(dataT, facF, otherF = NULL, aggF = NULL, DV, subF,
  aovv = NULL, tDec = 2, pDec = 3, pCut = 0.001, dDec = 3)
}
\arguments{
\item{dataT}{The data table}

\item{facF}{The name of the column containing the factor on which the comparisons are operated.}

\item{otherF}{Any other factors in the ANOVA design. Useful for simple effects. Vector of column names or comma separated string of column names.}

\item{aggF}{Factors to aggregate on. Useful when decomposing main effects. Vector of column names or comma separated string of column names.}

\item{DV}{Name of the column containing the dependant variable.}

\item{subF}{Name of the column containing the subject code.}

\item{aovv}{(optional) aov object}

\item{tDec}{Number of decimals for the t statistic}

\item{pDec}{Number of decimals for the p value(s)}

\item{pCut}{Cutting point for the p value}

\item{dDec}{Number of decimals for Cohen's d}
}
\value{
Augmented data.table with the means, t tests and effect sizes.
}
\description{
Calculates paired comparisons for a factor on a data.table
}
\section{References}{
 
Lakens, D. (2013). Calculating and reporting effect sizes to facilitate cumulative science: a practical primer for t-tests and ANOVAs. Frontiers in Psychology, 4.
}

\examples{
data("donn_ag",package="muCoVL")
dataT<-donn_ag
facF<-"Trial_Block"
otherF<-"Group,dist"
aggF<-"dur"
DV<-"relE"
subF<-"sub"
pairedCVLDT(dataT,facF,otherF,aggF,DV,subF)
}
