% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{scriptDS}
\alias{scriptDS}
\title{Generate knitR script for descriptive statistics}
\usage{
scriptDS(dataT, facF, DV, subF, levelV = 2, file = "", alpha = 0.05)
}
\arguments{
\item{dataT}{The data table}

\item{facF}{The name of the column containing the factor on which the comparisons are operated.}

\item{DV}{Name of the column containing the dependant variable.}

\item{subF}{Name of the column containing the subject code.}

\item{levelV}{Heading level.A number between 2 and 5}

\item{file}{Name of the file to output to. If the file already exists, it will append _1 to the name.}

\item{alpha}{Significance level.}
}
\value{
Prints the knitR code in the specified file.
}
\description{
This helper function prints the knitR code necessary to print the descriptive stats for all the specified factors in a data.table.
}
\examples{
data("donn_ag")
dataT<-donn_ag
facF<-"Trial_Block,Group,dist,dur"
levelV<-2
file<-""
DV<-"relE"
subF<-"sub"
alpha<-.05

scriptDS(dataT = dataT,facF = facF,DV = DV,subF = subF,levelV = levelV,file=file)
}
