% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{tableAPA}
\alias{tableAPA}
\title{Create APA style tables}
\usage{
tableAPA(DataT, bys, DVcol, nameIV, nameslvl1, nameslvl2 = NULL,
  nameslvl3 = NULL, tablew = "100\%")
}
\arguments{
\item{DataT}{Data.table containing the data}

\item{bys}{vector of the names of the columns in DataT that will form the row names. Slowest to fastest varying columns.}

\item{DVcol}{string variable containing the name of the columns in DataT that contain the DVs. Single string of names separated by commas.}

\item{nameIV}{string variable containing the name of the independant Variable.}

\item{nameslvl1}{vector of string variables giving the label of each column. If two column label cells are merged, then put NA for the second column's label.}

\item{nameslvl2}{vector of string variables giving the label of each column at the second level of headings. If two column label cells are merged, then put NA for the second column's label. (optional)}

\item{nameslvl3}{vector of string variables giving the label of each column at the third level of headings. If two column label cells are merged, then put NA for the second column's label. (optional)}

\item{tablew}{width of the table, either in pixels.}
}
\value{
string variable containing the html code for the table.
}
\description{
This function formats a table in HTML format that is ready for publication.
}
