% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCoVL.R
\name{VL_unEvenPoly}
\alias{VL_unEvenPoly}
\title{Uneven polynomial contrasts}
\usage{
VL_unEvenPoly(w, scaled = T)
}
\arguments{
\item{w}{vector of weights for the contrasts}

\item{scaled}{Whether the output should be scaled or not (boolean).}
}
\value{
A matrix of polynomial contrasts.
}
\description{
Returns a matrix of polynomial contrasts in accordance with a set of weights.
}
